requirements.txt
Django==5.1.1
djangorestframework==3.15.2
googlemaps==4.10.0
pytesseract==0.3.13
pillow==10.4.0
requests==2.32.3
python-dotenv==0.19.1
django-cors-headers==4.0.0